+++ src/test/java/test/validation/Addressbook.java
@@ -7,6 +7,7 @@ import java.util.Iterator;
 import be.baur.sda.DataNode;
 import be.baur.sda.SDA;
 import be.baur.sds.SDS;
+import be.baur.sds.Schema;
 import be.baur.sds.validation.Error;
 import be.baur.sds.validation.ErrorList;
 import be.baur.sds.validation.Validator;
@@ -27,11 +28,15 @@ public final class Addressbook {
 		DataNode document = SDA.parse(new InputStreamReader(sda, "UTF-8"));
 
 		InputStream sds = Addressbook.class.getResourceAsStream("/addressbook.sds");
-		Validator validator = SDS.parse(new InputStreamReader(sds, "UTF-8")).newValidator();
+		Schema schema = SDS.parse(new InputStreamReader(sds, "UTF-8"));
+		Validator validator = schema.newValidator();
 
-		ErrorList errors = validator.validate(document, "contact");
+		validator.setTypeName("contact"); // set to existing type but not what we expect
+		ErrorList errors = validator.validate(document);
 		t.ts1("F01", errors.get(0) + "", "/addressbook: got 'addressbook', but 'contact' was expected");
-		errors = validator.validate(document, null);
+		
+		validator.setTypeName(null); // try again with any matching type
+		errors = validator.validate(document);
 		//for (Error error : errors) System.out.println(error.toString());
 		Iterator<Error> e = errors.iterator();
 		
@@ -51,7 +56,16 @@ public final class Addressbook {
 		t.ts1("F14", e.next() + "", "/addressbook/contact[6]/address/postalcode: got 'postalcode', but 'housenumber' was expected");
 		t.ts1("F15", e.next() + "", "/addressbook/contact[6]: content missing at end of 'contact'; expected 'phone' or 'email'");
 		t.ts1("F16", e.next() + "", "/addressbook/contact[7]/person: empty value not allowed; 'person' is not nullable");
-		t.ts1("F16", e.next() + "", "/addressbook/contact[7]/address: value 'nowhere' does not match pattern 'home|work'");
-		t.ts1("F99", e.hasNext() + "", "false");
+		t.ts1("F17", e.next() + "", "/addressbook/contact[7]/address: value 'nowhere' does not match pattern 'home|work'");
+		t.ts1("F18", e.hasNext() + "", "false");
+		
+		validator.setTypeName("contact"); // now validate an actual contact (the first one)
+		errors = validator.validate(document.get("contact"));
+		e = errors.iterator();
+
+		t.ts1("F19", e.next() + "", "/addressbook/contact[1]/person/about: only complex content is expected for node 'about'");
+		t.ts1("F20", e.next() + "", "/addressbook/contact[1]/address/housenumber: got 'housenumber', but 'streetname' or 'postalcode' was expected");
+		t.ts1("F21", e.next() + "", "/addressbook/contact[1]/email[2]: 'email' was not expected in 'contact'");
+		t.ts1("F22", e.hasNext() + "", "false");
 	}
 }